Add tests for CapabilityStatement Ref: https://github.com/Aidbox/aidbox-python-sdk/issues/69
