+++ Pipfile
@@ -8,6 +8,7 @@ aiohttp = "~=3.10.2"
 sqlalchemy = "~=2.0.5"
 fhirpy = "~=1.3.0"
 jsonschema = "~=4.17.3"
+fhirpathpy = "==2.0.0"
 
 [dev-packages]
 pytest-aiohttp = "~=1.0.4"
@@ -20,8 +21,6 @@ autohooks-plugin-ruff = "~=23.11.0"
 autohooks-plugin-black = "~=23.7.0"
 coloredlogs = "*"
 ruff = "*"
-
-# dependencies for python 3.8
 exceptiongroup = "~=1.2.1"
 tomli = "~=2.0.1"
 async-timeout = "~=4.0.3"
